models/training/requirements.txt
transformers>=4.48.0
datasets>=2.14.0
evaluate>=0.4.0
rouge_score>=0.1.2
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.27.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tensorboard>=2.13.0
tqdm>=4.65.0
tokenizers>=0.15.0
